feat: Restructure analytics feature by subdomains
- Create subdomain folders: credentials/, insights/, post-analytics/, tag-analytics/, summaries/
- Move API type definitions to respective subdomain folders
- Create subdomain-specific API handlers and operations
- Update main API file to aggregate handlers from all subdomains
- Split AnalyticsHandlers type into subdomain-specific handler types
- Maintain same external IPC interface for backward compatibility
- Update all import paths to reflect new structure
- Add type safety with subdomain-specific handler interfaces

Closes #29
