--- src/features/analytics/api-type/index.ts
@@ -1,6 +1,0 @@
-export * from "./fansly-credentials";
-export * from "./handlers";
-export * from "./insights";
-export * from "./post-analytics";
-export * from "./summary-analytics";
-export * from "./tag-analytics";
