+++ src/features/analytics/credentials/api.ts
@@ -1,0 +1,8 @@
+import { updateFanslyCredentialsFromFetch } from "./operations";
+import { CredentialsHandlers } from "./api-type";
+
+export const credentialsHandlers: CredentialsHandlers = {
+  updateFanslyCredentialsFromFetch: async (_: unknown, fetchRequest: string) => {
+    return updateFanslyCredentialsFromFetch(fetchRequest);
+  },
+};
