+++ src/features/analytics/tag-analytics/api-type.ts
@@ -23,3 +23,18 @@ export type TagCorrelationData = {
   combinedPerformance: number;
   coOccurrenceCount: number;
 };
+
+export type TagAnalyticsHandlers = {
+  getTagAnalytics: (_: unknown, params: TagAnalyticsParams) => Promise<TagPerformanceMetrics[]>;
+  getTagPerformanceMetrics: (
+    _: unknown,
+    tagIds: number[],
+    timeRange: { start: Date; end: Date }
+  ) => Promise<TagPerformanceMetrics[]>;
+  getTagCorrelations: (_: unknown, dimensionId?: number) => Promise<TagCorrelationData[]>;
+  getTagTrends: (
+    _: unknown,
+    tagIds: number[],
+    timeRange: { start: Date; end: Date }
+  ) => Promise<any>;
+};
